+++ changelog.md
@@ -1,6 +1,7 @@
 ### Revision History
 * 4.0.21-SNAPSHOT
   * Updated `JsonHttpProxy` constructor to take `org.apache.http.HttpHost` parameter instead of individual hostname, port, scheme parameters in order to construct the `HttpHost`
+  * Bug fix: sys.branch.permissions not being created when creating branch in sys boot version.
 * 4.0.20
   * Bug fix: Updated `NCube.convertExistingAxisToRefAxis()` to not throw NPE when axis and axis to reference do not have any overlapping columns
   * `NCube` stack traces, when listing the coordinate, only show the first 'n' characters of the value (default 1000 - can be overridden in `application.properties`).