+++ changelog.md
@@ -7,6 +7,7 @@
   * Bug fix: `sys.branch.permissions` not being created when creating branch in sys boot version.
   * Bug fix: A new column would lose any metaproperties associated with it on merge / update.
   * Bug fix: Custom sys.menu would not open properly when viewing a released version.
+  * Bug fix: Read default Spring configuration properties correctly
 * 4.0.20
   * Bug fix: Updated `NCube.convertExistingAxisToRefAxis()` to not throw NPE when axis and axis to reference do not have any overlapping columns
   * `NCube` stack traces, when listing the coordinate, only show the first 'n' characters of the value (default 1000 - can be overridden in `application.properties`).